backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
stripe==9.9.0
redis==5.0.4
apscheduler==3.10.4
httpx==0.27.0
python-multipart==0.0.9
Pillow==10.3.0
psycopg2-binary==2.9.9
greenlet==3.0.3
anyio==4.3.0
aiosmtplib==3.0.1
